+++ src/screens/Standby.tsx
@@ -1,30 +1,32 @@
 import { motion } from "framer-motion";
-import type { AnimTypes } from "./animations";
-import { getAnimations, sectionVariants } from "./animations";
-import { Casters } from "./components/Casters";
-import { CurrentMapStats } from "./components/CurrentMapStats";
-import { FooterContent } from "./components/FooterContent";
-import { GameplaySvgMask } from "./components/GameplayTransparencyMask";
-import { HeaderContent } from "./components/HeaderContent";
-import { Logo } from "./components/Logo";
-import { MainContent } from "./components/MainContent";
-import { PlayerInfo } from "./components/PlayerInfo";
-import { ScoreBars } from "./components/ScoreBars";
+import type { AnimTypes } from "~/animations";
+import { getAnimations, sectionVariants } from "~/animations";
+import { Casters } from "~/components/Casters";
+import { Chat } from "~/components/Chat";
+import { FooterContent } from "~/components/FooterContent";
+import { GameplaySvgMask } from "~/components/GameplayTransparencyMask";
+import { HeaderContent } from "~/components/HeaderContent";
+import { Logo } from "~/components/Logo";
+import { MainContent } from "~/components/MainContent";
+import { PlayerInfo } from "~/components/PlayerInfo";
+import { StageInfo } from "~/components/StageInfo";
 
-interface VersusScreenProps {
+interface StandbyScreenProps {
   from?: string;
   to: string;
 }
 
-export function VersusScreen({ from, to }: VersusScreenProps) {
+export function StandbyScreen({ from, to }: StandbyScreenProps) {
   // treat this component as self (to) and other as from
   const anims: AnimTypes = getAnimations(to, from ?? "");
 
   const slideDirection: 1 | -1 = 1;
+
   return (
     <div>
       <GameplaySvgMask />
       <div id="main" className="no-background">
+        {/* Header */}
         <motion.div
           key={`header-${to}`}
           {...(anims.header === "slide"
@@ -36,11 +38,15 @@ export function VersusScreen({ from, to }: VersusScreenProps) {
           <HeaderContent>
             <div id="top">
               <PlayerInfo playerNum={1} />
-              <ScoreBars />
+              <div id="middle">
+                <StageInfo />
+              </div>
               <PlayerInfo playerNum={2} />
             </div>
           </HeaderContent>
         </motion.div>
+
+        {/* Main */}
         <motion.div
           key={`main-${to}`}
           {...(anims.main === "slide"
@@ -53,6 +59,8 @@ export function VersusScreen({ from, to }: VersusScreenProps) {
             <div id="gameplay"></div>
           </MainContent>
         </motion.div>
+
+        {/* Footer */}
         <motion.div
           key={`footer-${to}`}
           {...(anims.footer === "slide"
@@ -65,7 +73,7 @@ export function VersusScreen({ from, to }: VersusScreenProps) {
             <div id="orange-line"></div>
             <div id="bottom">
               <Logo />
-              <CurrentMapStats />
+              <Chat />
               <Casters />
             </div>
           </FooterContent>
