+++ src/screens/Versus.tsx
@@ -1,32 +1,30 @@
 import { motion } from "framer-motion";
-import type { AnimTypes } from "./animations";
-import { getAnimations, sectionVariants } from "./animations";
-import { Casters } from "./components/Casters";
-import { Chat } from "./components/Chat";
-import { FooterContent } from "./components/FooterContent";
-import { GameplaySvgMask } from "./components/GameplayTransparencyMask";
-import { HeaderContent } from "./components/HeaderContent";
-import { Logo } from "./components/Logo";
-import { MainContent } from "./components/MainContent";
-import { PlayerInfo } from "./components/PlayerInfo";
-import { StageInfo } from "./components/StageInfo";
+import type { AnimTypes } from "~/animations";
+import { getAnimations, sectionVariants } from "~/animations";
+import { Casters } from "~/components/Casters";
+import { CurrentMapStats } from "~/components/CurrentMapStats";
+import { FooterContent } from "~/components/FooterContent";
+import { GameplaySvgMask } from "~/components/GameplayTransparencyMask";
+import { HeaderContent } from "~/components/HeaderContent";
+import { Logo } from "~/components/Logo";
+import { MainContent } from "~/components/MainContent";
+import { PlayerInfo } from "~/components/PlayerInfo";
+import { ScoreBars } from "~/components/ScoreBars";
 
-interface StandbyScreenProps {
+interface VersusScreenProps {
   from?: string;
   to: string;
 }
 
-export function StandbyScreen({ from, to }: StandbyScreenProps) {
+export function VersusScreen({ from, to }: VersusScreenProps) {
   // treat this component as self (to) and other as from
   const anims: AnimTypes = getAnimations(to, from ?? "");
 
   const slideDirection: 1 | -1 = 1;
-
   return (
     <div>
       <GameplaySvgMask />
       <div id="main" className="no-background">
-        {/* Header */}
         <motion.div
           key={`header-${to}`}
           {...(anims.header === "slide"
@@ -38,15 +36,11 @@ export function StandbyScreen({ from, to }: StandbyScreenProps) {
           <HeaderContent>
             <div id="top">
               <PlayerInfo playerNum={1} />
-              <div id="middle">
-                <StageInfo />
-              </div>
+              <ScoreBars />
               <PlayerInfo playerNum={2} />
             </div>
           </HeaderContent>
         </motion.div>
-
-        {/* Main */}
         <motion.div
           key={`main-${to}`}
           {...(anims.main === "slide"
@@ -59,8 +53,6 @@ export function StandbyScreen({ from, to }: StandbyScreenProps) {
             <div id="gameplay"></div>
           </MainContent>
         </motion.div>
-
-        {/* Footer */}
         <motion.div
           key={`footer-${to}`}
           {...(anims.footer === "slide"
@@ -73,7 +65,7 @@ export function StandbyScreen({ from, to }: StandbyScreenProps) {
             <div id="orange-line"></div>
             <div id="bottom">
               <Logo />
-              <Chat />
+              <CurrentMapStats />
               <Casters />
             </div>
           </FooterContent>
