+++ tsconfig.json
@@ -23,7 +23,7 @@
 
     "baseUrl": ".",
     "paths": {
-      "@/*": ["./src/*"]
+      "~/*": ["./src/*"]
     },
 
     // Some stricter flags (disabled by default)
